Reword confusing sentence
It's actually irrelevant if this code is called from the previous
example or not; the point is that we want to test this library's public
method and it calls a bunch of private methods in a bunch of different
cases. It's more of a contrasting example to the previous code, and less
a building upon the previous example, which the original sentence led me
to expect.