Remove "conversely" from constructions that are not converse.
In many cases, the actual intent behind the colloquial misuses of the
word "conversely" don't fit cleanly into the actual definitions of
"converse", "inverse", and "contrapositive".  This commit fixes the
malapropisms that most easily fit into these categories, uses similar
constructions such as "likewise" and "on the other hand" where
appropriate, and tries to avoid the construction altogether in others.

The one instance of "conversely" left untouched seems to be close enough
to its correct usage.